[IMP] finish basic support of API
Adds the findr route.
If there are more than one attendee, does as many queries to get lists
of preferences and aggregate the results.
The ranking is pretty basic at this point (the TODO stays :-/).
Also we don't weigh preferences if there are more than one attendee
with the same preferences.
A future version could include a scoring system, also in the results.
The output restaurant validation is pretty light at this point.
There is also logging to monitor failings of the current code.